+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -13,6 +13,9 @@ Release History
 * Allow `network watcher flow-log show/configure` to be used with classic NSGs.
 * Add `network watcher run-configuration-diagnostic` command.
 * Fix `network watcher test-connectivity` command and add `--method`, `--valid-status-codes` and `--headers` properties.
+* `network express-route create/update`: Add `--allow-global-reach` flag.
+* `network vnet subnet create/update`: Add support for `--delegation`.
+* Added `network vnet subnet list-available-delegations` command.
 
 2.2.4
 +++++